move tree declaration to core package
