+++ src/main/java/com/github/sybila/huctl/package.kt
@@ -1,5 +1,6 @@
 package com.github.sybila.huctl
 
+import com.github.sybila.Unary
 import org.antlr.v4.runtime.tree.ParseTree
 import org.antlr.v4.runtime.tree.ParseTreeProperty
